+++ disjunct_knn_noop.go
